"Hiding" poll_until in a mixin; exluding it from the public API.
(cherry picked from commit 8660964aae48b085d33a521310e3551292d864c8)
