+++ test/unit/ae_page_objects_test.rb
@@ -41,24 +41,4 @@ class AePageObjectsTest < AePageObjectsTestCase
     end
     assert_equal "Time appears to be frozen", raised.message
   end
-
-  def test_poll_until__with_mocks
-    AePageObjects.expects(:wait_until).yields
-    Capybara.expects(:using_wait_time).with(0).yields
-
-    block = mock
-    block.expects(:called).times(1)
-
-    AePageObjects.poll_until do
-      block.called
-    end
-  end
-
-  def test_poll_until__without_mocks
-    result = AePageObjects.poll_until do
-      :hello
-    end
-
-    assert_equal :hello, result
-  end
 end
